Fix _stream_helper
The stream helper that yields the streamed status for push, pull and
build operations was reading data in chunks of 4096 bytes but didn't
support buffering and thus didn't behave well when more than 4096 bytes
were received at once from the socket.

This implementation uses makefile() to get a file object from the socket
on which we can use readline() so we can read line by line the
chunked-encoding data streamed to us by the Docker daemon.

Signed-off-by: Maxime Petazzoni <max@signalfuse.com>