Added support for the same DOCKER_HOST environment variable the official Docker client uses
For an example case where this is useful, imagine I'm using an application which uses docker-py, but doesn't offer a flag for me to control the base_url it tries to connect to.  If I run my Docker on a separate host or on a non-standard socket path, then without this change I have to actually modify the source code of the application I'm using.  With this change, everything can "just work" if I set the DOCKER_HOST variable correctly, just like the Docker client itself does. :)