Universally-compatible reading of chunked streams
Docker introduced newlines in stream output in version 0.9
(https://github.com/dotcloud/docker/pull/4276), but not to all
endpoints - POST /images/create, for example, does not include them.

This reverts to the old, less pleasant implementation of
_stream_helper(), with a manual check for newlines to fix the problem
described in #176 and fixed in #184, without the accompanying
regression. It should work against Docker 0.8, 0.9 and 0.10, both when
building and when pulling.