Store the registry key the same way of resolving.
Before this change, if the configuration file have a authentication
configuration for `docker.example.com`, it won't be successfully used
for pulling `docker.transloc.com/example:latest`. This is because the
configuration is looked up by the key `https://docker.example.com/v1/`.
This change fixed this bug to store the configurations also in this full
format.