Added Content-Type header to pull procedure
I did a packet capture between a manual docker pull with swarm and a docker pull from the python client with swarm. The python client was missing a header 'Content-Type' that was causing an API error with swarm. When I added this header to the pull procedure, the python client started working with swarm.