More rigorous support for .dockerignore
Fixes #581. The current implementation for parsing .dockerignore files
and applying the exclusion to the build path does not account for many
common and uncommon syntax and behaviors (see `tests/test.py` for a
few of these cases). The problem is that the rules for parsing
.dockerignore files are complex. Luckily,
[pathspec](https://github.com/cpburnz/python-path-specification)
provides a robust implementation of .gitignore rules, which are almost
identical for .dockerignore.

This PR uses the `pathspec` library in place of the current
`fnmatch_any`. Additional tests for .dockerignore have been added. All
tests pass for py27.