Fix log driver integration test
By using `json-file` as the use case we can run this without needing
to be on a machine with syslog installed.

Even if syslog was installed, this test was still failing as the
`log_opt` of 'key1' is an invalid option.

This test is slightly different style to the others, it hopefully
includes an example of how to space out tests for increased readability,
we construct our tests in 3 phases(generally) and include a space imbetween
them:

- setup

- perform action

- assertions

Expanded variable names is important for readability/clarity. Also
reduced the number of assertions to focus on the logic we're testing.

Signed-off-by: Mazz Mosley <mazz@houseofmnowster.com>