Remove validation of supported log drivers
By having this hardcoded list of log drivers, it is a bottleneck
to us supporting more log drivers.

The daemon already validates if a log driver is valid or not, so rather
than duplicating that validation, let's pass the log_driver along.

This allows support for new/more log drivers as they become supported
in docker without having to wait for both docker-py and docker-compose
to support them.

Keeping the current list of log driver types for backwards compatibility.

Signed-off-by: Mazz Mosley <mazz@houseofmnowster.com>