create_host_config requires version
When network_mode was introduced to have a default of 'default'
it broke backwards compatibility for API versions < 1.20

To fix this, we need to check a version, to do that, we need the
version passed in. To ensure this fix doesn't break backwards
compatibility itself, we use the DEFAULT_DOCKER_API_VERSION as
a default if version isn't passed in.

Signed-off-by: Mazz Mosley <mazz@houseofmnowster.com>