Make volume binds tests work on any host
Instead of creating the test directory directly on the host, create it
by starting a container with the directory bind-mounted, so that it
doesn't matter whether the daemon is local, in a VM or remote.

This removes the need to make /tmp a volume in the test container, and
to share it with the dind container.

Signed-off-by: Aanand Prasad <aanand.prasad@gmail.com>