Allow a host to be passed for the IP portion of a port binding.
Previously, if a hostname were provided in place of an IP address in a port
binding definition, the docker daemon would simply ignore it and bind to all
interfaces instead of a specific one.  This change makes it so that a hostname
can be used in place of the IP.  The address for that host will be resolved
to an IP and then the IP is passed to the docker daemon.  This still requires
that the IP that it resolves to exists on the docker host.

This is particularly useful in situations where it is common to move containers
from one docker host to another.  One could have a docker-compose.yml, for
instance, that does not need to change (using the hostname in the 'ports'
definition).  When the container is to be moved to a new host, DNS can just be
updated to reflect the address on the new host to be used.  Then docker-compose
will use the right address when it's brought up on that host.

Signed-off-by: Tommy Beadle <tbeadle@gmail.com>