Pass file object to Tarfile.addfile()
This resolves an issue where TarFile.gettarinfo() doesn't include the
file object, meaning that TarFile.addfile(TarFile.gettarinfo()) doesn't
pass the test suite. Instead, this uses an open() within a try...except
block to include a file object for each file without passing a file
object when the path is a directory.

Signed-off-by: Christian Bundy <christianbundy@fraction.io>