Ensure default timeout is used by API Client
The `from_env` method on the `docker` module passed `None` as the
value for the `timeout` keyword argument which overrode the default
value in the initialiser, taken from `constants` module.

This sets the default in the initialiser to `None` and adds logic
to set that, in the same way that `version` is handled.

Signed-off-by: grahamlyons <graham@grahamlyons.com>