Fix plugin model upgrade
When upgrading a plugin via the model interface, it would yield the
following error:

    AttributeError: 'Plugin' object has no attribute '_reload'

It appears that the proper method is `self.reload()`. This is what is
used by all other methods in the class and base. I'm not finding any
references to `_reload` apart from this instance in the project either.

I've verified that this patch fixes the issue on my machine and all
tests pass.

Signed-off-by: Ian Fijolek <ian@iamthefij.com>