Add return types to all __str__ methods
If this library is to ship type hints, it is important that those
type hints are valid.
A type checker (e.g. `mypy` or `pyright`) can check that those type
hints are valid, but to do so, the codebase must be type hinted throughout.
This PR is one small step towards that goal.

This change replaces all instances of `def __str__(self):` with
`def __str__(self) -> str:`.
I chose to do a small PR, as to add passing type checking to the
whole codebase involves hundreds or maybe thousands of changes,
some of which may need real consideration.