Remove and return container when start fails
When the container fail to start and the user has set the `remove` flag
we should remove the containter, also if the flag is not set, we lose
track of the container, leaving it dangling on the system. By adding
this execption return whenever the container fails to start we provide
means to the user to remove the container if needed.

Signed-off-by: Filipe Utzig <filipe@gringolito.com>