fix waiting for auto_remove container
use the same logic as the cli to wait for the exit of a container when
running one: ensure that a container run with auto_remove set to True
has been removed when the function returns.

this prevents a race condition when trying to run another container with
the same name directly afterwards.

signed-off-by: hugues de keyzer <docker@hugues.info>